+++ whitelist-authenticator/src/main/java/de/intension/authentication/schools/CacheStatus.java
@@ -1,0 +1,10 @@
+package de.intension.authentication.schools;
+
+/**
+ * Represents different states of @{@link WhiteListCache}
+ */
+public enum CacheStatus {
+    NOT_INITIALIZED,
+    OUTDATED,
+    VALID
+}